#10 bottomSheetView 코드 수정
